Add custom watches for styles + translations
The logging of npm-watch is quite heavy.
Especially when an exception ocurrs, the
whole npm log is showing up.
