+++ dev/hmr.js
@@ -1,5 +1,5 @@
-require("chokidar-socket-emitter")({
-  port: require("../config.json").port.hmr,
+module.exports = require("chokidar-socket-emitter")({
+  port: require("../server/config.json").port.hmr,
   path: "src",
   chokidar: {
     ignored: /___jb.*$/ // IntelliJ Idea creates and removes a temporary file on each save
