+++ dev/livereload.js
@@ -1,5 +1,5 @@
 require("livereload")
   .createServer({
-    port: require("../config.json").port.livereload
+    port: require("../server/config.json").port.livereload
   })
   .watch([".tmp/main.css", "src/dev/index.dev.html"]);
