+++ dev/utils.js
@@ -1,0 +1,25 @@
+var fs = require("fs");
+var _ = require("lodash");
+var mkdirp = require("mkdirp");
+var glob = require("glob");
+var path = require("path");
+
+module.exports = {
+  getFiles (src) {
+    return new Promise((resolve, reject) =>
+      glob(src, (e, files) =>
+        e ? reject(e) : resolve(files)))
+      .then(files => _.reject(files, f => _.includes(f, "*")));
+  },
+  readFile (src) {
+    return new Promise((resolve, reject) =>
+      fs.readFile(src, "utf-8", (e, content) =>
+        e ? reject(e) : resolve(content)));
+  },
+  writeFile (dest, content, options) {
+    return new Promise((resolve, reject) =>
+      mkdirp(path.dirname(dest), e =>
+        e ? reject(e) : fs.writeFile(dest, content, options || {}, e =>
+          e ? reject(e) : resolve())));
+  }
+};
